move stream and connection interfaces to abc
