+++ libp2p/network/stream/net_stream.py
@@ -1,5 +1,4 @@
-from libp2p.stream_muxer.muxed_connection_interface import IMuxedConn
-from libp2p.stream_muxer.muxed_stream_interface import IMuxedStream
+from libp2p.stream_muxer.abc import IMuxedConn, IMuxedStream
 
 from .net_stream_interface import INetStream
