+++ libp2p/network/typing.py
@@ -1,4 +1,4 @@
 from typing import Awaitable, Callable
-from libp2p.stream_muxer.muxed_stream_interface import IMuxedStream
+from libp2p.stream_muxer.abc import IMuxedStream
 
 GenericProtocolHandlerFn = Callable[[IMuxedStream], Awaitable[None]]
