+++ libp2p/pubsub/pubsub_notifee.py
@@ -5,7 +5,7 @@ from multiaddr import Multiaddr
 from libp2p.network.network_interface import INetwork
 from libp2p.network.notifee_interface import INotifee
 from libp2p.network.stream.net_stream_interface import INetStream
-from libp2p.stream_muxer.muxed_connection_interface import IMuxedConn
+from libp2p.stream_muxer.abc import IMuxedConn
 
 if TYPE_CHECKING:
     import asyncio
