+++ libp2p/stream_muxer/mplex/mplex_stream.py
@@ -1,6 +1,5 @@
 import asyncio
-from libp2p.stream_muxer.muxed_stream_interface import IMuxedStream
-from libp2p.stream_muxer.muxed_connection_interface import IMuxedConn
+from libp2p.stream_muxer.abc import IMuxedStream, IMuxedConn
 
 
 from .constants import HeaderTags
